+++ R/reexport_generics.R
@@ -1,4 +1,13 @@
-#Re-export the tidier generics
-#' @importFrom generics tidy 
-#' @export 
+#' @importFrom generics tidy
+#' @export
+#' @seealso [tidy.rmcorr()]
 generics::tidy
+
+#' @importFrom generics glance
+#' @export
+#' @seealso [glance.rmcorr()]
+generics::glance
+
+#' @importFrom generics required_pkgs
+#' @export
+generics::required_pkgs